requirements.txt
google-api-python-client==2.100.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.22.0
beautifulsoup4==4.12.2